requirements.txt
appnope==0.1.3
asgiref==3.7.2
asttokens==2.4.0
attrs==23.1.0
backcall==0.2.0
black==23.10.0
blessed==1.20.0
bpython==0.24
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
curtsies==0.4.2
cwcwidth
decorator==5.1.1
Django==4.2.6
django-debug-toolbar==4.2.0
django-extensions==3.2.3
django-rest-framework==0.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-spectacular==0.26.5
executing==2.0.0
flake8==6.1.0
greenlet==3.0.0
idna==3.4
inflection==0.5.1
ipython==8.16.1
jedi==0.19.1
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
matplotlib-inline==0.1.6
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==23.2
parso==0.8.3
pathspec==0.11.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.1.0
platformdirs==3.11.0
prompt-toolkit==3.0.39
psycopg2-binary
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.11.1
pyflakes==3.1.0
Pygments==2.16.1
PyJWT==2.8.0
python-dotenv==1.0.0
pytz==2023.3.post1
pyxdg==0.28
PyYAML==6.0.1
referencing==0.30.2
requests==2.31.0
rpds-py==0.10.6
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.3
traitlets==5.11.2
uritemplate==4.1.1
urllib3==2.0.7
wcwidth==0.2.8
